quote-viewer/requirements.txt
os
cs50
flask
flask_session
requests
werkzeug.security
urllib.parse
functools
